+++ psh/parser/recursive_descent/__init__.py
@@ -11,11 +11,8 @@ from .parser import Parser
 
 __all__ = [
     'Parser',
-    'ContextBaseParser',
     'ParserContext',
     'ParserProfiler',
-    'HeredocInfo',
     'ParseError',
     'ErrorContext',
-    'TokenGroups',
 ]